requirements/local.txt
django-debug-toolbar
django-extensions
mixer==6.0.1
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2